requirements.txt
aioredis==1.3.1
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
backcall==0.2.0
cffi==1.14.3
channels==3.0.2
channels-redis==3.2.0
constantly==15.1.0
cryptography==3.2.1
daphne==3.0.1
decorator==4.4.2
Django==2.2
django-model-utils==4.0.0
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
ipdb==0.13.4
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
msgpack==1.0.0
parso==0.7.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.2
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pytz==2020.4
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
traitlets==4.3.3
Twisted==20.3.0
txaio==20.4.1
wcwidth==0.2.5
zope.interface==5.2.0
